Removed uneccessary type casting. Equations should now be easier to read

+++ amplitudePhaseCharts.cpp
@@ -99,8 +99,8 @@ QPointF AmplitudePhaseCalculation::getIntertionFirstOrder(DataAcquired_t& data,
 {
     QPointF result;
 
-    result.setX( data.k / ( pow((qreal)data.t1, 2) * pow(omega, 2) + (qreal)1 ) );
-    result.setY( (-data.k * (qreal)data.t1 * omega)/( pow((qreal)data.t1, 2) * pow(omega, 2) + 1 ) );
+    result.setX( data.k / ( pow(data.t1, (qreal)2) * pow(omega, (qreal)2) + (qreal)1 ) );
+    result.setY( (-data.k * data.t1 * omega)/( pow(data.t1, (qreal)2) * pow(omega, (qreal)2) + (qreal)1 ) );
 
     return result;
 }
@@ -113,10 +113,14 @@ QPointF AmplitudePhaseCalculation::getIntertionSecondOrder(DataAcquired_t& data,
 {
     QPointF result;
 
-    result.setX( data.k * ( (qreal)1 - pow(omega, 2) * (qreal)data.t1 * (qreal)data.t2 ) /
-                 ( (qreal)1 + pow(( omega * (qreal)data.t1 ), 2) + pow(( omega * (qreal)data.t2 ), 2) + pow( pow(omega, 2) * (qreal)data.t1 * (qreal)data.t2, 2) ) );
-    result.setY( -data.k * omega * ( (qreal)data.t1 + (qreal)data.t2 ) /
-                 ( (qreal)1 + pow(( omega * (qreal)data.t1 ), 2) + pow(( omega * (qreal)data.t2 ), 2) + pow( pow(omega, 2) * (qreal)data.t1 * (qreal)data.t2, 2) ) );
+    result.setX( data.k * ( (qreal)1 - pow(omega, (qreal)2) * data.t1 * data.t2 ) /
+                 ( (qreal)1 + pow(( omega * data.t1 ), (qreal)2) +
+                   pow(( omega * data.t2 ), (qreal)2) +
+                   pow( pow(omega, (qreal)2) * data.t1 * data.t2, (qreal)2) ) );
+    result.setY( -data.k * omega * ( data.t1 + data.t2 ) /
+                 ( (qreal)1 + pow(( omega * data.t1 ), (qreal)2)
+                   + pow(( omega * data.t2 ), (qreal)2) +
+                   pow( pow(omega, (qreal)2) * data.t1 * data.t2, (qreal)2) ) );
 
 
     return result;
@@ -131,25 +135,24 @@ QPointF AmplitudePhaseCalculation::getIntertionThirdOrder(DataAcquired_t& data,
     QPointF result;
 
     result.setX(
-    data.k * ( ( (qreal)1 - pow(omega, 2) * (qreal)data.t1 * (qreal)data.t2 ) -
-    (pow(omega, 2) * (qreal)data.t1 * (qreal)data.t3) -
-    (pow(omega, 2) * (qreal)data.t1 * (qreal)data.t2) )
-    /
-    ( ( (qreal)1 + pow(omega, 2) * pow((qreal)data.t1, 2) ) *
-    ( (qreal)1 + pow(omega, 2) * pow((qreal)data.t2, 2) ) *
-    ( (qreal)1 + pow(omega, 2) * pow((qreal)data.t3, 2) ) )
-    );
+                data.k * ( ( (qreal)1 - pow(omega, (qreal)2) * data.t1 * data.t2 ) -
+                           (pow(omega, (qreal)2) * data.t1 * data.t3) -
+                           (pow(omega, (qreal)2) * data.t1 * data.t2) )
+                /
+                ( ( (qreal)1 + pow(omega, (qreal)2) * pow(data.t1, (qreal)2) ) *
+                  ( (qreal)1 + pow(omega, (qreal)2) * pow(data.t2, (qreal)2) ) *
+                  ( (qreal)1 + pow(omega, (qreal)2) * pow(data.t3, (qreal)2) ) )
+                );
 
     result.setY(
-    data.k * ( (-omega * (qreal)data.t3) -
-    (omega * (qreal)data.t1) -
-    (omega * (qreal)data.t2) +
-    ( pow(omega,3) * (qreal)data.t1 * (qreal)data.t2 * (qreal)data.t3) )
-    /
-    ( ( (qreal)1 + pow(omega, 2) * pow((qreal)data.t1, 2) ) *
-    ( (qreal)1 + pow(omega, 2) * pow((qreal)data.t2, 2) ) *
-    ( (qreal)1 + pow(omega, 2) * pow((qreal)data.t3, 2) ) )
-    );
+                data.k * ( (-omega * data.t3) -
+                           (omega * data.t1) -
+                           (omega * data.t2) +
+                           ( pow(omega, (qreal)3) * data.t1 * data.t2 * data.t3) ) /
+                ( ( (qreal)1 + pow(omega, (qreal)2) * pow(data.t1, (qreal)2) ) *
+                  ( (qreal)1 + pow(omega, (qreal)2) * pow(data.t2, (qreal)2) ) *
+                  ( (qreal)1 + pow(omega, (qreal)2) * pow(data.t3, (qreal)2) ) )
+                );
 
     return result;
 }
@@ -163,36 +166,36 @@ QPointF AmplitudePhaseCalculation::getIntertionFourthOrder(DataAcquired_t& data,
     QPointF result;
 
     result.setX(
-    data.k * ( ( (qreal)1 - pow(omega, 2) * (qreal)data.t3 * (qreal)data.t4 ) -
-    (pow(omega, 2) * (qreal)data.t1 * (qreal)data.t4) -
-    (pow(omega, 2) * (qreal)data.t1 * (qreal)data.t2) -
-    (pow(omega, 2) * (qreal)data.t2 * (qreal)data.t4) -
-    (pow(omega, 2) * (qreal)data.t2 * (qreal)data.t3) +
-    (pow(omega, 4) * (qreal)data.t1 * (qreal)data.t2 * (qreal)data.t3 * (qreal)data.t4 )
-    )
-    /
-    ( ( (qreal)1 + pow(omega, 2) * pow((qreal)data.t1, 2) ) *
-    ( (qreal)1 + pow(omega, 2) * pow((qreal)data.t2, 2) ) *
-    ( (qreal)1 + pow(omega, 2) * pow((qreal)data.t3, 2) ) *
-    ( (qreal)1 + pow(omega, 2) * pow((qreal)data.t4, 2) ) )
-    );
+                data.k * ( ( (qreal)1 - pow(omega, (qreal)2) * data.t3 * data.t4 ) -
+                           (pow(omega, (qreal)2) * data.t1 * data.t4) -
+                           (pow(omega, (qreal)2) * data.t1 * data.t2) -
+                           (pow(omega, (qreal)2) * data.t2 * data.t4) -
+                           (pow(omega, (qreal)2) * data.t2 * data.t3) +
+                           (pow(omega, (qreal)4) * data.t1 * data.t2 * data.t3 * data.t4 )
+                           )
+                /
+                ( ( (qreal)1 + pow(omega, (qreal)2) * pow(data.t1, (qreal)2) ) *
+                  ( (qreal)1 + pow(omega, (qreal)2) * pow(data.t2, (qreal)2) ) *
+                  ( (qreal)1 + pow(omega, (qreal)2) * pow(data.t3, (qreal)2) ) *
+                  ( (qreal)1 + pow(omega, (qreal)2) * pow(data.t4, (qreal)2) ) )
+                );
 
     result.setY(
-     data.k * ( (-omega * (qreal)data.t4) -
-    (omega * (qreal)data.t3) -
-    (omega * (qreal)data.t1) +
-    ( pow(omega,3) * (qreal)data.t1 * (qreal)data.t2 * (qreal)data.t4) -
-    (omega * (qreal)data.t2) +
-    (pow(omega, 3) * (qreal)data.t2 * (qreal)data.t3 * (qreal)data.t4) +
-    (pow(omega, 3) * (qreal)data.t1 * (qreal)data.t2 * (qreal)data.t4) +
-    (pow(omega, 3) * (qreal)data.t1 * (qreal)data.t2 * (qreal)data.t3)
-    )
-    /
-    ( ( (qreal)1 + pow(omega, 2) * pow((qreal)data.t1, 2) ) *
-    ( (qreal)1 + pow(omega, 2) * pow((qreal)data.t2, 2) ) *
-    ( (qreal)1 + pow(omega, 2) * pow((qreal)data.t3, 2) ) *
-    ( (qreal)1 + pow(omega, 2) * pow((qreal)data.t4, 2) ) )
-    );
+                data.k * ( (-omega * data.t4) -
+                           (omega * data.t3) -
+                           (omega * data.t1) +
+                           ( pow(omega, (qreal)3) * data.t1 * data.t2 * data.t4) -
+                           (omega * data.t2) +
+                           (pow(omega, (qreal)3) * data.t2 * data.t3 * data.t4) +
+                           (pow(omega, (qreal)3) * data.t1 * data.t2 * data.t4) +
+                           (pow(omega, (qreal)3) * data.t1 * data.t2 * data.t3)
+                           )
+                /
+                ( ( (qreal)1 + pow(omega, (qreal)2) * pow(data.t1, (qreal)2) ) *
+                  ( (qreal)1 + pow(omega, (qreal)2) * pow(data.t2, (qreal)2) ) *
+                  ( (qreal)1 + pow(omega, (qreal)2) * pow(data.t3, (qreal)2) ) *
+                  ( (qreal)1 + pow(omega, (qreal)2) * pow(data.t4, (qreal)2) ) )
+                );
 
     return result;
 }
